Enable networking for nucleo_f756zg
Networking is not tested!

FYI: Coredumps are not functional in this, however, because the data at
`s_ram_backed_coredump_region` is being erased on reboot 😕:

```plaintext
(gdb) x /32c s_ram_backed_coredump_region
0x20012dc0 <s_ram_backed_coredump_region>:      0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'
0x20012dc8 <s_ram_backed_coredump_region+8>:    0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'        0 '\000'
0x20012dd0 <s_ram_backed_coredump_region+16>:   0 '\000'        0 '\000'        0 '\000'        0 '\000'        76 'L'  0 '\000'        0 '\000'        0 '\000'
0x20012dd8 <s_ram_backed_coredump_region+24>:   1 '\001'        -128 '\200'     0 '\000'        0 '\000'        32 ' '  26 '\032'       1 '\001'        32 ' '
```

The first 4 bytes should be `"CORE"`.

Could be related to the ROM bootloader, see this document:

https://www.st.com/resource/en/application_note/cd00167594-stm32-microcontroller-system-memory-boot-mode-stmicroelectronics.pdf

Might need to move the `s_ram_backed_coredump_region` to a different
location.